feat: add anti aliasing post processing (#90 via #103)
* docs: add missing parameter description

* feat: add anti aliasing demo scene

* misc: add logging message

* fix: regression bug intensity at overpoint

not sure how that slipped through

* refactor: pull up pixel ray origin calculation

allows the anti aliasing camera to create multiple rays per pixel in different sub-pixel locations

* refactor: procedural camera uses new camera api

* refactor: timing wrapper camera uses new api

* feat: create super sample camera implementation

* feat: implement anti aliasing demo scene

* feat: create canvas resampler

* test: create canvas resampler tests

* misc: identify possible bug

* fix: use correct indices

* test: add rotated grid tests

* feat: add fixed pixel sampling patterns

* misc: add more verbose exception message

* misc: identify bug

* test: add dynamic sampler offset generator test

* feat: implement dynamic super sampler mode

* feat: add no aa mode for comparison

* misc: use default aa mode in aa demo scene

* misc: move unused import

* test: add end to end aa test

a triangle covers half the camera diagonally, the triangle edge can be inspected to evaluate the performance of the sample mode

* fix: convenience method dispatches to base method

* fix: correct expected test output

* docs: add method javadoc

* fix: convert focus camera fov to radians

+++ src/main/java/com/BudgiePanic/rendering/reporting/TimingWrapper.java
@@ -41,10 +41,10 @@ public class TimingWrapper implements Camera {
     public int height() { return internal.height(); }
 
     @Override
-    public Ray createRay(int pixelColumn, int pixelRow, float time) { return internal.createRay(pixelColumn, pixelRow, time); }
+    public Ray createRay(float pixelColumn, float pixelRow, float time) { return internal.createRay(pixelColumn, pixelRow, time); }
 
     @Override
-    public Color pixelAt(World world, int pixelColumn, int pixelRow, float time) {
+    public Color pixelAt(World world, float pixelColumn, float pixelRow, float time) {
         return internal.pixelAt(world, pixelColumn, pixelRow, time);
     }
